module3/upload_data/pyproject.toml
[project]
name = "upload-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-storage>=3.9.0",
    "pyarrow>=23.0.0",
]
